Only delete config file if a config file name wasn't provided

+++ tests/e2e/lib-compat.sh
@@ -43,8 +43,10 @@ deploy_stackrox_with_roxie_compat() {
         retrying_kubectl create ns "${namespace}" </dev/null
     fi
 
+    local cleanup_config_file="false"
     if [[ -z "$config_file" ]]; then
         config_file="$(mktemp)"
+        cleanup_config_file="true"
     elif [[ ! -e "$config_file" ]]; then
         touch "$config_file"
     fi
@@ -68,7 +70,10 @@ EOF
     roxie_config_from_environment_compat "$config_file" "$namespace"
 
     deploy_stackrox_with_roxie "$config_file"
-    rm -f "$config_file"
+
+    if [[ "$cleanup_config_file" == "true" ]]; then
+        rm -f "$config_file"
+    fi
 }
 
 # This function translates environment settings into a roxie configuration.
